+++ src/search.rs
@@ -271,7 +271,8 @@ pub fn search(command_params: CommandParams) {
                             sketch_params.use_aa,
                             command_params.est_ci,
                             command_params.detailed_out,
-                            !*fw
+                            !*fw,
+                            command_params.short_header,
                         );
                         if *fw == true{
                             *fw = false;
@@ -294,7 +295,8 @@ pub fn search(command_params: CommandParams) {
         sketch_params.use_aa,
         command_params.est_ci,
         command_params.detailed_out,
-        !*first_write.lock().unwrap()
+        !*first_write.lock().unwrap(),
+        command_params.short_header,
     );
     info!("Searching time: {}", now.elapsed().as_secs_f32());
 }